modified attendance routing

+++ database/seeders/TestDataSeeder.php
@@ -34,7 +34,7 @@ class TestDataSeeder extends Seeder
             }
         }
 /////////////////////////////////////////////
-               
+
         $items = [
             [
                 'id'             => 1,
@@ -51,7 +51,7 @@ class TestDataSeeder extends Seeder
                 'location'    => 'MLA Hostel',
                 'administrative_office_id' => 2,
             ],
-            
+
         ];
 
         foreach ($items as $item) {
@@ -117,7 +117,7 @@ class TestDataSeeder extends Seeder
                 'is_controlling_officer' => 1,
                 'level' =>65,
             ],
-            
+
         ];
 
         foreach ($items as $item) {
@@ -126,14 +126,14 @@ class TestDataSeeder extends Seeder
             }
         }
         ////////////////
-   
+
         $items = [
             [
                 'id'     => 1,
                 'srismt'  => 'Sri',
                 'name' => 'Vishnu MM',
                 'aadhaarid' => '123451',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
@@ -142,7 +142,7 @@ class TestDataSeeder extends Seeder
                 'srismt'  => 'Smt',
                 'name' => 'Mujeeba SI',
                 'aadhaarid' => '123452',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
@@ -151,7 +151,7 @@ class TestDataSeeder extends Seeder
                 'srismt'  => 'Sri',
                 'name' => 'Hari',
                 'aadhaarid' => '123422',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
@@ -160,7 +160,7 @@ class TestDataSeeder extends Seeder
                 'srismt'  => 'Sri',
                 'name' => 'Sreeja',
                 'aadhaarid' => '123423',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
@@ -169,7 +169,7 @@ class TestDataSeeder extends Seeder
                 'srismt'  => 'Smt',
                 'name' => 'Nazia',
                 'aadhaarid' => '223423',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
@@ -178,7 +178,7 @@ class TestDataSeeder extends Seeder
                 'srismt'  => 'Sri',
                 'name' => 'Sachi',
                 'aadhaarid' => '323423',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
@@ -187,7 +187,7 @@ class TestDataSeeder extends Seeder
                 'srismt'  => 'Sri',
                 'name' => 'Abdul Nsar',
                 'aadhaarid' => '123523',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
@@ -196,7 +196,7 @@ class TestDataSeeder extends Seeder
                 'srismt'  => 'Sri',
                 'name' => 'Hari P',
                 'aadhaarid' => '153523',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
@@ -206,67 +206,67 @@ class TestDataSeeder extends Seeder
                 'srismt'  => 'Sri',
                 'name' => 'Shaji C Baby',
                 'aadhaarid' => '153423',
-                'employee_type'=> 'permanent',
+
                 'status' => 'active',
                 'is_shift' => 0,
             ],
-            
-            
+
+
         ];
         foreach ($items as $item) {
             $emp = Employee::where('aadhaarid',$item['aadhaarid'])->first();
-            $emp->delete();
-        
+            if($emp) $emp->delete();
+
                 Employee::insert($item);
-        
+
         }
       ///////////////////////////////////////
-         
+
       $items = [
         [
             'id'             => 1,
             'name'    => 'IT',
             'short_code'    => 'it',
-            'seat_of_controling_officer_id' => 3,
+            'seat_of_controlling_officer_id' => 3,
             'seat_of_reporting_officer_id' => 1,
             'office_location_id' => 2,
-            
+
         ],
         [
             'id'             => 2,
             'name'    => 'IT(eNiyamasabha)',
             'short_code'    => 'it.en',
-            'seat_of_controling_officer_id' => 3,
+            'seat_of_controlling_officer_id' => 3,
             'seat_of_reporting_officer_id' => 2,
             'office_location_id' => 2,
         ],
         [
             'id'             => 3,
             'name'    => 'Office of DS(eNiyamasabha)',
-            'short_code'    => 'oo.ds.it',
-            'seat_of_controling_officer_id' => 4,
+          //  'short_code'    => 'oo.ds.it',
+            'seat_of_controlling_officer_id' => 4,
             'seat_of_reporting_officer_id' => 4,
             'office_location_id' => 2,
         ],
         [
             'id'             => 4,
             'name'    => 'Office of HariP',
-            'short_code'    => 'oo.harip',
-            'js_as_ss_employee_id'    => 6,
-            'seat_of_controling_officer_id' => 5,
+           // 'short_code'    => 'oo.harip',
+            'js_as_ss_employee_id'    => 8,
+            'seat_of_controlling_officer_id' => 5,
             'seat_of_reporting_officer_id' => 5,
             'office_location_id' => 2,
         ],
         [
             'id'             => 4,
             'name'    => 'Office of Secretary',
-            'short_code'    => 'oo.secretary',
-            'js_as_ss_employee_id'    => 7, //
-            'seat_of_controling_officer_id' => 5,
+           // 'short_code'    => 'oo.secretary',
+            'js_as_ss_employee_id'    => 9, //
+            'seat_of_controlling_officer_id' => 5,
             'seat_of_reporting_officer_id' => 5,
             'office_location_id' => 2,
         ],
-        
+
     ];
     foreach ($items as $item) {
         if (!Section::find($item['id'])) {
